% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_3d_plot.R
\name{interactive3dplot}
\alias{interactive3dplot}
\title{Interactively plotting 3D attainment plots}
\usage{
interactive3dplot(
  x,
  percentiles = c(0, 50, 100),
  col = c("black", "white"),
  type = "point",
  psize = 3,
  xlabel = "X",
  ylabel = "Y",
  zlabel = "Z"
)
}
\arguments{
\item{x}{Either a matrix of data values, or a data frame, or a list of data frames of exactly three columns.}

\item{percentiles}{\code{(numeric())} - Vector indicating which percentile should be plot. The default is to plot only the median attainment curve. If \code{type='cube'} then only the first percentile will be used.}

\item{col}{A vector of colours used to plot.}

\item{type}{\code{(character(1))} - string giving the type of plot desired. The following values are possible, \code{'scatter'} and \code{'cube'}.}

\item{psize}{Size of plotted points.}

\item{xlabel, ylabel, zlabel}{Axes labels.}
}
\value{
A 3D plot of the attainment points and surfaces
}
\description{
Interactively plotting 3D attainment plots
}
\examples{
options("warn"=-1)
library(interactiveeafplots)

extdata_dir <- system.file(package="interactiveeafplots", "extdata")
x <- moocore::read_datasets(file.path(extdata_dir, "spherical-250-10-3d.txt"))
# Select only the last 3 datasets
x <- x[x[, 4] >= 8, ]

interactive3dplot(x, c(0,50,100), col=c("red","orange","yellow"),
                  type="scatter", psize=3, xlabel="Objective 1",
                  ylabel="Objective 2", zlabel="Objective 3")

# Will take a long time to render.
# Opens in separate window, so run by yourself.
# interactive3dplot(x, c(0,50,100), col=c("red","yellow"), type="cube",
#                   psize=3, xlabel="Objective 1", ylabel="Objective 2",
#                   zlabel="Objective 3")

}
