% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_symdev_plot.R
\name{interactivesymdevplot}
\alias{interactivesymdevplot}
\title{Interactively plotting the Symmetric Deviation function}
\usage{
interactivesymdevplot(
  x,
  percentiles = c(0, 50, 100),
  maximise = FALSE,
  col = c("white", "black"),
  threshold = NULL,
  thrcol = "cyan",
  type = "point",
  lty = "solid",
  psize = 0.1,
  pshape = 16,
  legend.pos = "topright",
  xaxis.side = "bottom",
  yaxis.side = "left",
  axes = TRUE,
  xlabel = NULL,
  ylabel = NULL,
  sci.notation = FALSE,
  extraLegend = "VE",
  plot = "plotly"
)
}
\arguments{
\item{x}{Either a matrix of data values, or a data frame, or a list of data frames of exactly three columns.}

\item{percentiles}{\code{(numeric())} - Vector indicating which percentile should be plot. The default is to plot only the median attainment curve.}

\item{maximise}{\code{(logical() | logical(1))} - Whether the objectives must be maximised instead of minimised. Either a single logical value that applies to all objectives or a vector of logical values, with one value per objective.}

\item{col}{A vector of colours used to plot.}

\item{threshold}{Vorob'ev threshold, e.g., as returned by \code{\link[moocore::vorobT()]{moocore::vorobT()}}.}

\item{thrcol}{Colour of Vorob'ev expectation}

\item{type}{\code{(character(1))} - string giving the type of plot desired. The following values are possible, \code{'point'} and \code{'area'}.}

\item{lty}{Line type. See \code{\link[plot.default()]{plot.default()}}.}

\item{psize}{Size of plotted points}

\item{pshape}{Shape of plotted point. See \code{\link[plot.default()]{plot.default()}}.}

\item{legend.pos}{The position of the legend, valid values are \code{'topright'}, \code{'top'}, \code{'topleft'}, \code{'right'}, \code{'left'}, \code{'bottom'}, \code{'bottomright'}, \code{'bottomleft'} and \code{'center'}. A value of \code{'none'} hides the legend.}

\item{xaxis.side}{On which side that xaxis is drawn. Valid values are \code{'below'} and \code{'above'}.}

\item{yaxis.side}{On which side that yaxis is drawn. Valid values are \code{'left'} and \code{'right'}.}

\item{axes}{A logical value indicating whether both axes should be drawn on the plot.}

\item{xlabel, ylabel}{Axes labels}

\item{sci.notation}{Generate scientific labels}

\item{extraLegend}{Legend label for Vorob'ev expectation.}

\item{plot}{Interactive or none interactive plot. Valid values are \code{'plotly'} and \code{'ggplot'}.}
}
\value{
A plot of the attainment surfaces of the symmetric deviation function.
}
\description{
Interactively plotting the Symmetric Deviation function
}
\examples{
options("warn"=-1)
suppressMessages(library(dplyr))
suppressMessages(library(plotly))
library(moocore)
library(stringr)

data(CPFs, package = "moocore")
res <- moocore::vorobT(CPFs, reference = c(2, 200))
print(res$threshold)
#> [1] 44.14062

# Display Vorob'ev expectation and attainment function
interactivesymdevplot(x=CPFs, threshold=res$threshold, col =c("red","yellow"),
                      percentiles = c(0,20,40,60,80,100), type="area",
                      thrcol="cyan", plot = "plotly", extraLegend="VE",
                      sci.notation = TRUE)

}
